bug de precio-peso solucionado

+++ _weight/dvr_HX712.c
@@ -61,14 +61,14 @@ void ReadHX712(void){ //by ERH
 	
 	//LCD_GLASS_Float(ADcode_pre, 0, LCD_TOTAL); 
    
-	if(ADcode_pre == 262143){
+	if(ADcode_pre == (unsigned long)262143){
 //	  srFlagScale.bErrorReadAdc=1;
-	}else if(ADcode_pre > 262143){
-		ADcode_pre = 524288 - ADcode_pre;
-		ADcode_pre = 32500 - ADcode_pre;
+	}else if(ADcode_pre > (unsigned long)262143){
+		ADcode_pre = (unsigned long)524288 - ADcode_pre;
+		ADcode_pre = (unsigned long)32500 - ADcode_pre;
 	//	srFlagScale.bErrorReadAdc = 0;
 	}else{
-		ADcode_pre += 32500;
+		ADcode_pre += (unsigned long)32500;
 //		srFlagScale.bErrorReadAdc = 0;
 	}
 	//LCD_GLASS_Float(ADcode_pre, 0, LCD_TOTAL);
